# Conflicts:
#	app/src/main/java/com/example/fit_i/ui/main/matching/MatchingFragment.kt

+++ app/src/main/java/com/example/fit_i/data/model/response/GetMCResponse.kt
@@ -8,8 +8,8 @@ data class GetMCResponse(
     val result: List<Result> ) {
 
     data class Result(
-        val matchingId: Int,
-        val trainerId: Int,
+        val matchingId: Long,
+        val trainerId: Long,
         val name: String,
         val profile: String,
         val school: String,
